Fix Spelling errors place container inside components
