+++ src/components/container/PriceFilter.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { changeFilter } from '../redux/crypto/cryptoActions';
-import css from '../styles/nav.module.css';
+import { changeFilter } from '../../redux/crypto/cryptoActions';
+import css from '../../styles/nav.module.css';
 
 const PriceFilter = ({ changeFilter }) => {
   const priceRanges = ['MORE THAN 10$', 'MORE THAN 50$', 'MORE THAN 100$', 'MORE THAN 200$', 'MORE THAN 1000$'];
